+++ Install/Install.ps1
@@ -511,7 +511,7 @@ $InstallationEntry = Add-PnPListItem -List "Installasjonslogg" -Values $InstallE
 
 ## Attempting to attach the log file to installation entry
 if ($null -ne $InstallationEntry) {
-    $File = Get-Item -Path $LogFilePath
+    $File = Get-Item -Path $LogFilePath -ErrorAction Continue
     if ($null -ne $File -and $File.Length -gt 0) {
         Write-Host "[INFO] Attaching installation log file to installation entry"
         $AttachmentOutput = Add-PnPListItemAttachment -List "Installasjonslogg" -Identity $InstallationEntry.Id -Path $LogFilePath -ErrorAction Continue